Drop dependency on pytest
Closes #49.
