+++ Makefile
@@ -4,6 +4,7 @@
 REPO_NAME:=$(shell basename `git rev-parse --show-toplevel`)
 VENV_NAME:='venv/$(REPO_NAME)'
 DEV_DEPENDS:='requirements/dev'
+PYTHON:='python'
 
 .DEFAULT_GOAL:= help
 .PHONY: help
@@ -14,7 +15,6 @@ help:
 	@echo '    setup           Setup for development'
 	@echo '    local-install   Install locally'
 	@echo '    local-uninstall Uninstall locally'
-	@echo '    update-depends  Re-compile requirements for development'
 	@echo '    lint            Re-lint by black and isort with setup.cfg'
 	@echo '    test            Run unittests'
 	@echo '    deploy          Release to PyPI server'
@@ -26,24 +26,20 @@ venv:
 	test -d $@ || mkdir -m 755 ./$@
 
 $(VENV_NAME): venv
-	test -d $(VENV_NAME) || python -m venv $(VENV_NAME)
+	test -d $(VENV_NAME) || $(PYTHON) -m venv $(VENV_NAME)
 	test -d $(VENV_NAME) || exit 1 ;
 
 .PHONY: setup
 setup: $(VENV_NAME)
-	$(VENV_NAME)/bin/python -m pip install -r $(DEV_DEPENDS).txt
+	$(VENV_NAME)/bin/python -m pip install --index-url https://pypi.org/simple -r $(DEV_DEPENDS)
 
 .PHONY: local-install
 local-install: $(VENV_NAME)
-	$(VENV_NAME)/bin/python -m pip install -e .
+	$(VENV_NAME)/bin/python -m pip install --index-url https://pypi.org/simple -e .
 
 .PHONY: local-uninstall
 local-uninstall:
 	$(VENV_NAME)/bin/python -m pip uninstall -y pip-licenses
-
-.PHONY: update-depends
-update-depends:
-	$(VENV_NAME)/bin/python -m pip-compile --extra dev -o requirements/dev.txt -U pyproject.toml
 
 .PHONY: lint
 lint:
@@ -54,7 +50,7 @@ lint:
 
 .PHONY: test
 test:
-	$(VENV_NAME)/bin/python -m pytest
+	$(VENV_NAME)/bin/python -m unittest discover --start-directory tests --verbose
 
 .PHONY: deploy
 deploy: build
@@ -76,9 +72,8 @@ clean:
 full-clean:: local-uninstall clean
 	rm -vrf *.egg-info 2>/dev/null || : ;
 	rm -vfrd ./{piplicenses,tests}/__pycache__ 2>/dev/null || : ;
-	rm -vfrd ./.coverage 2>/dev/null || : ;
+	rm -vfrd ./.coverage* 2>/dev/null || : ;
 	rm -vfrd ./.mypy_cache 2>/dev/null || : ;
-	rm -vfrd ./.pytest_cache 2>/dev/null || : ;
 
 .PHONY: un-setup
 un-setup:: full-clean
