+++ requirements/dev.txt
@@ -1,190 +1,20 @@
-#
-# This file is autogenerated by pip-compile with Python 3.9
-# by the following command:
-#
-#    pip-compile --extra=dev --output-file=requirements/dev.txt pyproject.toml
-#
-attrs==25.4.0
-    # via flake8-bugbear
-autopep8==2.3.2
-    # via pip-licenses-cli (pyproject.toml)
-backports-tarfile==1.2.0
-    # via jaraco-context
 black==25.9.0
-    # via pip-licenses-cli (pyproject.toml)
 boolean-py==5.0
-    # via license-expression
-build==1.3.0
-    # via pip-tools
-certifi==2025.10.5
-    # via requests
-cffi==2.0.0
-    # via cryptography
-charset-normalizer==3.4.4
-    # via requests
-click==8.1.8
-    # via
-    #   black
-    #   pip-tools
 coverage[toml]==7.10.7
-    # via pytest-cov
 cryptography==46.0.2
-    # via pip-licenses-cli (pyproject.toml)
 docutils==0.22.2
-    # via
-    #   pip-licenses-cli (pyproject.toml)
-    #   readme-renderer
-exceptiongroup==1.3.0
-    # via pytest
 flake8==7.3.0
-    # via
-    #   flake8-bugbear
-    #   pep8-naming
-    #   pip-licenses-cli (pyproject.toml)
 flake8-bugbear==24.12.12
-    # via pip-licenses-cli (pyproject.toml)
-id==1.5.0
-    # via twine
-idna==3.11
-    # via requests
-importlib-metadata==8.7.0
-    # via
-    #   build
-    #   keyring
-    #   twine
-iniconfig==2.1.0
-    # via pytest
 isort==6.1.0
-    # via pip-licenses-cli (pyproject.toml)
-jaraco-classes==3.4.0
-    # via keyring
-jaraco-context==6.0.1
-    # via keyring
-jaraco-functools==4.3.0
-    # via keyring
-keyring==25.6.0
-    # via twine
 license-expression==30.4.4
-    # via pip-licenses-cli (pyproject.toml)
-markdown-it-py==3.0.0
-    # via rich
-mccabe==0.7.0
-    # via flake8
-mdurl==0.1.2
-    # via markdown-it-py
-more-itertools==10.8.0
-    # via
-    #   jaraco-classes
-    #   jaraco-functools
 mypy==1.18.2
-    # via pip-licenses-cli (pyproject.toml)
 mypy-extensions==1.1.0
-    # via
-    #   black
-    #   mypy
-nh3==0.3.1
-    # via readme-renderer
-packaging==25.0
-    # via
-    #   black
-    #   build
-    #   pytest
-    #   twine
-pathspec==0.12.1
-    # via
-    #   black
-    #   mypy
 pep8-naming==0.15.1
-    # via pip-licenses-cli (pyproject.toml)
 pip-licenses-lib==0.6.0
-    # via pip-licenses-cli (pyproject.toml)
-pip-tools==7.5.1
-    # via pip-licenses-cli (pyproject.toml)
-platformdirs==4.4.0
-    # via black
-pluggy==1.6.0
-    # via pytest
 prettytable==3.16.0
-    # via pip-licenses-cli (pyproject.toml)
-pycodestyle==2.14.0
-    # via
-    #   autopep8
-    #   flake8
-pycparser==2.23
-    # via cffi
-pyflakes==3.4.0
-    # via flake8
-pygments==2.19.2
-    # via
-    #   readme-renderer
-    #   rich
-pypandoc==1.15
-    # via pip-licenses-cli (pyproject.toml)
-pyproject-hooks==1.2.0
-    # via
-    #   build
-    #   pip-tools
-pytest==8.4.2
-    # via
-    #   pytest-cov
-pytest-cov==7.0.0
-    # via pip-licenses-cli (pyproject.toml)
-readme-renderer==44.0
-    # via twine
-requests==2.32.5
-    # via
-    #   id
-    #   requests-toolbelt
-    #   twine
-requests-toolbelt==1.0.0
-    # via twine
-rfc3986==2.0.0
-    # via twine
-rich==14.2.0
-    # via twine
+setuptools==80.9.0
 tomli==2.3.0 ; python_version < "3.11"
-    # via
-    #   autopep8
-    #   black
-    #   build
-    #   coverage
-    #   mypy
-    #   pip-licenses-cli (pyproject.toml)
-    #   pip-tools
-    #   pytest
 tomli-w==1.2.0
-    # via pip-licenses-cli (pyproject.toml)
 twine==6.2.0
-    # via pip-licenses-cli (pyproject.toml)
-types-colorama==0.4.15.20250801
-    # via pip-licenses-cli (pyproject.toml)
 types-docutils==0.22.2.20251006
-    # via
-    #   pip-licenses-cli (pyproject.toml)
-    #   types-pygments
-types-pexpect==4.9.0.20250916
-    # via pip-licenses-cli (pyproject.toml)
-types-pygments==2.19.0.20250809
-    # via pip-licenses-cli (pyproject.toml)
-typing-extensions==4.15.0
-    # via
-    #   black
-    #   exceptiongroup
-    #   mypy
-    #   rich
-urllib3==2.5.0
-    # via
-    #   requests
-    #   twine
-wcwidth==0.2.14
-    # via prettytable
 wheel==0.45.1
-    # via
-    #   pip-licenses-cli (pyproject.toml)
-    #   pip-tools
-zipp==3.23.0
-    # via importlib-metadata
-
-# The following packages are considered to be unsafe in a requirements file:
-# pip
-# setuptools
