+++ requirements/main.txt
@@ -1,9 +1,3 @@
-#
-# This file is autogenerated by pip-compile with Python 3.9
-# by the following command:
-#
-#    pip-compile --output-file=requirements/main.txt pyproject.toml
-#
 boolean-py==5.0
     # via license-expression
 license-expression==30.4.4
